Moved generic part of blended container into separate class
